feat(auth): add refresh token rotation (dual-cookie access/refresh) and fix IDOR(check if it's a real user) on user update/delete

+++ back-end/src/user/users.controller.ts
@@ -1,8 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, ParseUUIDPipe, /*ValidationPipe, UsePipes*/ } from '@nestjs/common'; // import des décorateurs utiles à UsersController
+import { Controller, Get, Post, Body, Patch, Param, Delete, ParseUUIDPipe, ForbiddenException, Req, /*ValidationPipe, UsePipes*/ } from '@nestjs/common'; // import des décorateurs utiles à UsersController
 import { UsersService } from './users.service'; // import de la definition de la classe UserService de users.service
 import { UpdateUserDto, extractUserUpdate } from './dto/update-user.dto'; // import de la classe UpdateUserDto
 import { UseGuards } from '@nestjs/common';
 import { JwtGuard } from '../common/guards/jwt.guard';
+import { Request } from 'express';
 
 // @UsePipes( new ValidationPipe() ) // instancie ValidationPipe pour qu'il check les regles du DTO lors d'une requete (actuellement instancié dans le main)
 @UseGuards( JwtGuard ) // applique le guard 'JwtGuard'
@@ -30,16 +31,22 @@ export class UsersController
 	///
 
 	@Patch( ':id' ) // méthode HTTP PATCH avec un arg (id) a récupérer avec @param
-	update( @Param( 'id', ParseUUIDPipe ) id: string, @Body() dto: UpdateUserDto ) // prends 2 params: id -> param recupéré sur url et DTO qui est instancié avec toutes la data du body de la requete
+	update( @Req() request, @Param( 'id', ParseUUIDPipe ) id: string, @Body() dto: UpdateUserDto ) // prends 2 params: id -> param recupéré sur url et DTO qui est instancié avec toutes la data du body de la requete
 	{
+		if ( request.user.id !== id )
+			throw new ForbiddenException( 'you can only modify your own account' );
+
 		return( this.usersService.update( id, extractUserUpdate( dto ))); // retourne le resultat de update de usersService -> l'objet complet user qui a été modifié
 	}
 
 	///
 
 	@Delete( ':id' ) // associe la méthode HTTP DELETE sur /users/:id à la méthode remove()
-	remove( @Param( 'id', ParseUUIDPipe ) id: string ) // récupère l'id du user dans l'url
+	remove( @Req() request, @Param( 'id', ParseUUIDPipe ) id: string ) // récupère l'id du user dans l'url
 	{
+		if ( request.user.id !== id )
+			throw new ForbiddenException( 'you can only delete your own account' );
+
 		return( this.usersService.remove( id ));
 	}
 }
